+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## <next>
 * Added babel-polyfill for DEMO
+* Added 2 default props : inputRef and onKeyDown
 
 ## 1.0.0
 * Initial release